Broke intcode functionality out of 2019/2
